MODIFY 12

+++ 12/7.c
@@ -2,11 +2,11 @@
 #include<math.h>
 #define PI 3.14159
 int main() {
-float r,c,s;
-    scanf("%lf\n",&r);
-    c=2*PI*r;
-        s=PI*r*r;
-    printf("%.3f\n",c);
- printf("%.3f\n",s);
+    double r, c, s;
+    scanf("%lf\n", &r);
+    c = 2 * PI * r;
+    s = PI * r * r;
+    printf("%.3lf\n", c);
+    printf("%.3lf\n", s);
     return 0;
 }
